Moving code from main.py
